+++ tests/com/yanncebron/m68kplugin/lang/psi/expression/M68kExpressionUtilTest.java
@@ -181,6 +181,18 @@ public class M68kExpressionUtilTest extends M68kPsiTestCase<M68kDcDirective> {
     assertEquals(6, computeConstantValue("2^4"));
   }
 
+  public void testComputeLogicalOrExpression() {
+    assertEquals(0, computeConstantValue("0||0"));
+    assertEquals(-1, computeConstantValue("1||1"));
+    assertEquals(-1, computeConstantValue("1||0"));
+  }
+
+  public void testComputeLogicalAndExpression() {
+    assertEquals(0, computeConstantValue("0&&0"));
+    assertEquals(0, computeConstantValue("0&&2"));
+    assertEquals(-1, computeConstantValue("1&&1"));
+  }
+
   private Object computeConstantValue(String expressionText) {
     return M68kExpressionUtil.getInstance().computeConstantValue(getExpression(expressionText));
   }